refactor(agent): split cmd/agent/main.go into per-subsystem files
Move-only stage 1 of catalog #9: peers, wshandler, recall, reconcile,
bootstrap, boxprobe, httpserve, sysmon, logging. run() and the wiring stay
in main.go (now ~1200 lines); moved blocks verified byte-identical.
